pyproject.toml
[project]
name = "flightboard-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]
